local defuns are not local
I was under the assumption that a local defun is part of the closure
of the enclosing defun. I was wrong. Local defuns are global bindings.
Thus the defun was replacing emacs's truncate with the local truncate.

This patch renames the function and makes it global,
in case it is ever needed.