feat: all refile categories support skip-refile flag
Modified all category --finalize functions (single-action, calendar,
delegate, habit, incubate) to check org-gtd-clarify--skip-refile flag
and call org-gtd-organize--update-in-place instead of org-gtd-refile--do
when the flag is set.

Added test to verify skip-refile behavior with single-action category
using spy to ensure org-refile is not called when flag is set.

This completes Task 7 of skip-refile implementation plan.